refactor(combo): add IgxComboBaseDirective
