+++ projects/igniteui-angular/src/lib/combo/combo.component.ts
@@ -1,7 +1,6 @@
 import { CommonModule } from '@angular/common';
 import {
-    AfterViewInit, ChangeDetectorRef, Component, ContentChild, ElementRef, EventEmitter, HostBinding, HostListener,
-    Input, NgModule, OnInit, OnDestroy, Output, TemplateRef, ViewChild, Optional, Inject, Injector
+    AfterViewInit, ChangeDetectorRef, Component, ElementRef, NgModule, OnInit, OnDestroy, Optional, Inject, Injector, ViewChild, Input
 } from '@angular/core';
 import {
     IgxComboItemDirective,
@@ -13,75 +12,28 @@ import {
     IgxComboToggleIconDirective,
     IgxComboClearIconDirective
 } from './combo.directives';
-import { FormsModule, ReactiveFormsModule, ControlValueAccessor, NG_VALUE_ACCESSOR, NgControl, AbstractControl } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule, ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { IgxCheckboxModule } from '../checkbox/checkbox.component';
 import { IgxSelectionAPIService } from '../core/selection';
-import { cloneArray, IBaseEventArgs, IBaseCancelableBrowserEventArgs, IBaseCancelableEventArgs, CancelableEventArgs } from '../core/utils';
+import { IBaseEventArgs, IBaseCancelableEventArgs, CancelableEventArgs } from '../core/utils';
 import { IgxStringFilteringOperand, IgxBooleanFilteringOperand } from '../data-operations/filtering-condition';
 import { FilteringLogic } from '../data-operations/filtering-expression.interface';
-import { IgxForOfModule, IForOfState, IgxForOfDirective } from '../directives/for-of/for_of.directive';
+import { IgxForOfModule, IForOfState } from '../directives/for-of/for_of.directive';
 import { IgxIconModule, IgxIconService } from '../icon/public_api';
 import { IgxRippleModule } from '../directives/ripple/ripple.directive';
 import { IgxToggleModule } from '../directives/toggle/toggle.directive';
 import { IgxButtonModule } from '../directives/button/button.directive';
-import { IgxDropDownModule, ISelectionEventArgs } from '../drop-down/public_api';
-import { IgxInputGroupModule, IgxInputGroupComponent } from '../input-group/input-group.component';
+import { IgxDropDownModule } from '../drop-down/public_api';
+import { IgxInputGroupModule } from '../input-group/input-group.component';
 import { IgxComboItemComponent } from './combo-item.component';
 import { IgxComboDropDownComponent } from './combo-dropdown.component';
 import { IgxComboFilteringPipe, IgxComboGroupingPipe } from './combo.pipes';
-import { OverlaySettings, AbsoluteScrollStrategy, AutoPositionStrategy } from '../services/public_api';
-import { noop, Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
-import { DisplayDensityBase, DisplayDensityToken, IDisplayDensityOptions } from '../core/density';
-import { IGX_COMBO_COMPONENT, IgxComboBase } from './combo.common';
+import { DisplayDensityToken, IDisplayDensityOptions } from '../core/density';
+import { IGX_COMBO_COMPONENT, IgxComboBaseDirective } from './combo.common';
 import { IgxComboAddItemComponent } from './combo-add-item.component';
 import { IgxComboAPIService } from './combo.api';
 import { EditorProvider } from '../core/edit-provider';
-import { IgxInputState, IgxInputDirective } from '../directives/input/input.directive';
 import { IgxInputGroupType, IGX_INPUT_GROUP_TYPE } from '../input-group/public_api';
-import { caseSensitive } from '@igniteui/material-icons-extended';
-import { SortingDirection } from '../data-operations/sorting-expression.interface';
-
-/**
- * @hidden
- */
-enum DataTypes {
-    EMPTY = 'empty',
-    PRIMITIVE = 'primitive',
-    COMPLEX = 'complex',
-    PRIMARYKEY = 'valueKey'
-}
-
-/**
- * @hidden
- */
-const ItemHeights = {
-    comfortable: 40,
-    cosy: 32,
-    compact: 28,
-};
-
-/**
- * @hidden
- * The default number of items that should be in the combo's
- * drop-down list if no `[itemsMaxHeight]` is specified
- */
-const itemsInContainer = 10;
-
-export enum IgxComboState {
-    /**
-     * Combo with initial state.
-     */
-    INITIAL = IgxInputState.INITIAL,
-    /**
-     * Combo with valid state.
-     */
-    VALID = IgxInputState.VALID,
-    /**
-     * Combo with invalid state.
-     */
-    INVALID = IgxInputState.INVALID
-}
 
 /** The filtering criteria to be applied on data search */
 export interface IComboFilteringOptions {
@@ -132,8 +84,6 @@ const diffInSets = (set1: Set<any>, set2: Set<any>): any[] => {
     return results;
 };
 
-let NEXT_ID = 0;
-
 /**
  *  Represents a drop-down list that provides editable functionalities, allowing users to choose an option from a predefined list.
  *
@@ -143,7 +93,7 @@ let NEXT_ID = 0;
  * @igxGroup Grids & Lists
  *
  * @remarks
- * It provides the ability to filter items as well as perform single or multiple selection with the provided data.
+ * It provides the ability to filter items as well as perform selection with the provided data.
  * Additionally, it exposes keyboard navigation and custom styling capabilities.
  * @example
  * ```html
@@ -162,492 +112,31 @@ let NEXT_ID = 0;
         { provide: NG_VALUE_ACCESSOR, useExisting: IgxComboComponent, multi: true }
     ]
 })
-export class IgxComboComponent extends DisplayDensityBase implements IgxComboBase, AfterViewInit, ControlValueAccessor, OnInit,
+export class IgxComboComponent extends IgxComboBaseDirective implements AfterViewInit, ControlValueAccessor, OnInit,
     OnDestroy, EditorProvider {
     /**
-     * Set custom overlay settings that control how the combo's list of items is displayed.
-     * Set:
-     * ```html
-     * <igx-combo [overlaySettings] = "customOverlaySettings"></igx-combo>
-     * ```
-     *
-     * ```typescript
-     *  const customSettings = { positionStrategy: { settings: { target: myTarget } } };
-     *  combo.overlaySettings = customSettings;
-     * ```
-     * Get any custom overlay settings used by the combo:
-     * ```typescript
-     *  const comboOverlaySettings: OverlaySettings = myCombo.overlaySettings;
-     * ```
-     */
-
-    @Input()
-    public overlaySettings: OverlaySettings = null;
-
-    /** @hidden @internal */
-    @ViewChild('inputGroup', { read: IgxInputGroupComponent, static: true }) public inputGroup: IgxInputGroupComponent;
-
-    /** @hidden @internal */
-    @ViewChild('comboInput', { read: IgxInputDirective, static: true }) public comboInput: IgxInputDirective;
-
-    /**
-     * @hidden @internal
-     */
-    @ViewChild(IgxComboDropDownComponent, { read: IgxComboDropDownComponent, static: true })
-    public dropdown: IgxComboDropDownComponent;
-
-    /**
-     * @hidden @internal
-     */
-    @ViewChild('searchInput')
-    public searchInput: ElementRef<HTMLInputElement> = null;
-
-    /**
-     * @hidden @internal
-     */
-    public get displaySearchInput(): boolean {
-        return this.filterable || this.allowCustomValues;
-    }
-
-    /**
-     * The custom template, if any, that should be used when rendering ITEMS in the combo list
-     *
-     * ```typescript
-     * // Set in typescript
-     * const myCustomTemplate: TemplateRef<any> = myComponent.customTemplate;
-     * myComponent.combo.itemTemplate = myCustomTemplate;
-     * ```
-     * ```html
-     * <!-- Set in markup -->
-     *  <igx-combo #combo>
-     *      ...
-     *      <ng-template igxComboItem>
-     *          <div class="custom-item" let-item let-key="valueKey">
-     *              <div class="custom-item__name">{{ item[key] }}</div>
-     *              <div class="custom-item__cost">{{ item.cost }}</div>
-     *          </div>
-     *      </ng-template>
-     *  </igx-combo>
-     * ```
-     */
-    @ContentChild(IgxComboItemDirective, { read: TemplateRef })
-    public itemTemplate: TemplateRef<any> = null;
-
-    /**
-     * The custom template, if any, that should be used when rendering the HEADER for the combo items list
-     *
-     * ```typescript
-     * // Set in typescript
-     * const myCustomTemplate: TemplateRef<any> = myComponent.customTemplate;
-     * myComponent.combo.headerTemplate = myCustomTemplate;
-     * ```
-     * ```html
-     * <!-- Set in markup -->
-     *  <igx-combo #combo>
-     *      ...
-     *      <ng-template igxComboHeader>
-     *          <div class="combo__header">
-     *              This is a custom header
-     *          </div>
-     *      </ng-template>
-     *  </igx-combo>
-     * ```
-     */
-    @ContentChild(IgxComboHeaderDirective, { read: TemplateRef })
-    public headerTemplate: TemplateRef<any> = null;
-
-    /**
-     * The custom template, if any, that should be used when rendering the FOOTER for the combo items list
-     *
-     * ```typescript
-     * // Set in typescript
-     * const myCustomTemplate: TemplateRef<any> = myComponent.customTemplate;
-     * myComponent.combo.footerTemplate = myCustomTemplate;
-     * ```
-     * ```html
-     * <!-- Set in markup -->
-     *  <igx-combo #combo>
-     *      ...
-     *      <ng-template igxComboFooter>
-     *          <div class="combo__footer">
-     *              This is a custom footer
-     *          </div>
-     *      </ng-template>
-     *  </igx-combo>
-     * ```
-     */
-    @ContentChild(IgxComboFooterDirective, { read: TemplateRef })
-    public footerTemplate: TemplateRef<any> = null;
-
-    /**
-     * The custom template, if any, that should be used when rendering HEADER ITEMS for groups in the combo list
-     *
-     * ```typescript
-     * // Set in typescript
-     * const myCustomTemplate: TemplateRef<any> = myComponent.customTemplate;
-     * myComponent.combo.headerItemTemplate = myCustomTemplate;
-     * ```
-     * ```html
-     * <!-- Set in markup -->
-     *  <igx-combo #combo>
-     *      ...
-     *      <ng-template igxComboHeaderItem let-item let-key="groupKey">
-     *          <div class="custom-item--group">Group header for {{ item[key] }}</div>
-     *      </ng-template>
-     *  </igx-combo>
-     * ```
-     */
-    @ContentChild(IgxComboHeaderItemDirective, { read: TemplateRef })
-    public headerItemTemplate: TemplateRef<any> = null;
-
-    /**
-     * The custom template, if any, that should be used when rendering the ADD BUTTON in the combo drop down
-     *
-     * ```typescript
-     * // Set in typescript
-     * const myCustomTemplate: TemplateRef<any> = myComponent.customTemplate;
-     * myComponent.combo.addItemTemplate = myCustomTemplate;
-     * ```
-     * ```html
-     * <!-- Set in markup -->
-     *  <igx-combo #combo>
-     *      ...
-     *      <ng-template igxComboAddItem>
-     *          <button class="combo__add-button">
-     *              Click to add item
-     *          </button>
-     *      </ng-template>
-     *  </igx-combo>
-     * ```
-     */
-    @ContentChild(IgxComboAddItemDirective, { read: TemplateRef })
-    public addItemTemplate: TemplateRef<any> = null;
-
-    /**
-     * The custom template, if any, that should be used when rendering the ADD BUTTON in the combo drop down
-     *
-     * ```typescript
-     * // Set in typescript
-     * const myCustomTemplate: TemplateRef<any> = myComponent.customTemplate;
-     * myComponent.combo.emptyTemplate = myCustomTemplate;
-     * ```
-     * ```html
-     * <!-- Set in markup -->
-     *  <igx-combo #combo>
-     *      ...
-     *      <ng-template igxComboEmpty>
-     *          <div class="combo--empty">
-     *              There are no items to display
-     *          </div>
-     *      </ng-template>
-     *  </igx-combo>
-     * ```
-     */
-    @ContentChild(IgxComboEmptyDirective, { read: TemplateRef })
-    public emptyTemplate: TemplateRef<any> = null;
-
-    /**
-     * The custom template, if any, that should be used when rendering the combo TOGGLE(open/close) button
-     *
-     * ```typescript
-     * // Set in typescript
-     * const myCustomTemplate: TemplateRef<any> = myComponent.customTemplate;
-     * myComponent.combo.toggleIconTemplate = myCustomTemplate;
-     * ```
-     * ```html
-     * <!-- Set in markup -->
-     *  <igx-combo #combo>
-     *      ...
-     *      <ng-template igxComboToggleIcon let-collapsed>
-     *          <igx-icon>{{ collapsed ? 'remove_circle' : 'remove_circle_outline'}}</igx-icon>
-     *      </ng-template>
-     *  </igx-combo>
-     * ```
-     */
-    @ContentChild(IgxComboToggleIconDirective, { read: TemplateRef })
-    public toggleIconTemplate: TemplateRef<any> = null;
-
-    /**
-     * The custom template, if any, that should be used when rendering the combo CLEAR button
-     *
-     * ```typescript
-     * // Set in typescript
-     * const myCustomTemplate: TemplateRef<any> = myComponent.customTemplate;
-     * myComponent.combo.clearIconTemplate = myCustomTemplate;
-     * ```
-     * ```html
-     * <!-- Set in markup -->
-     *  <igx-combo #combo>
-     *      ...
-     *      <ng-template igxComboClearIcon>
-     *          <igx-icon>clear</igx-icon>
-     *      </ng-template>
-     *  </igx-combo>
-     * ```
-     */
-    @ContentChild(IgxComboClearIconDirective, { read: TemplateRef })
-    public clearIconTemplate: TemplateRef<any> = null;
-
-    /**
-     * @hidden @internal
-     */
-    @ViewChild(IgxForOfDirective, { static: true })
-    public virtualScrollContainer: IgxForOfDirective<any>;
-
-    /**
-     * Emitted when item selection is changing, before the selection completes
-     *
-     * ```html
-     * <igx-combo (selectionChanging)='handleSelection()'></igx-combo>
-     * ```
-     */
-    @Output()
-    public selectionChanging = new EventEmitter<IComboSelectionChangingEventArgs | ISelectionEventArgs>();
-
-    /**
-     * Emitted before the dropdown is opened
-     *
-     * ```html
-     * <igx-combo opening='handleOpening($event)'></igx-combo>
-     * ```
-     */
-    @Output()
-    public opening = new EventEmitter<IBaseCancelableBrowserEventArgs>();
-
-    /**
-     * Emitted after the dropdown is opened
-     *
-     * ```html
-     * <igx-combo (opened)='handleOpened($event)'></igx-combo>
-     * ```
-     */
-    @Output()
-    public opened = new EventEmitter<IBaseEventArgs>();
-
-    /**
-     * Emitted before the dropdown is closed
-     *
-     * ```html
-     * <igx-combo (closing)='handleClosing($event)'></igx-combo>
-     * ```
-     */
-    @Output()
-    public closing = new EventEmitter<IBaseCancelableBrowserEventArgs>();
-
-    /**
-     * Emitted after the dropdown is closed
-     *
-     * ```html
-     * <igx-combo (closed)='handleClosed($event)'></igx-combo>
-     * ```
-     */
-    @Output()
-    public closed = new EventEmitter<IBaseEventArgs>();
-
-    /**
-     * Emitted when an item is being added to the data collection
-     *
-     * ```html
-     * <igx-combo (addition)='handleAdditionEvent($event)'></igx-combo>
-     * ```
-     */
-    @Output()
-    public addition = new EventEmitter<IComboItemAdditionEvent>();
-
-    /**
-     * Emitted when the value of the search input changes (e.g. typing, pasting, clear, etc.)
-     *
-     * ```html
-     * <igx-combo (searchInputUpdate)='handleSearchInputEvent($event)'></igx-combo>
-     * ```
-     */
-    @Output()
-    public searchInputUpdate = new EventEmitter<IComboSearchInputEventArgs>();
-
-    /**
-     * Emitted when new chunk of data is loaded from the virtualization
-     *
-     * ```html
-     * <igx-combo (dataPreLoad)='handleDataPreloadEvent($event)'></igx-combo>
-     * ```
-     */
-    @Output()
-    public dataPreLoad = new EventEmitter<IForOfState>();
-
-    /**
-     * Gets/gets combo id.
+     * Defines whether the caseSensitive icon should be shown in the search input
      *
      * ```typescript
      * // get
-     * let id = this.combo.id;
+     * let myComboShowSearchCaseIcon = this.combo.showSearchCaseIcon;
      * ```
      *
      * ```html
      * <!--set-->
-     * <igx-combo [id]='combo1'></igx-combo>
-     * ```
-     */
-    @HostBinding('attr.id')
-    @Input()
-    public id = `igx-combo-${NEXT_ID++}`;
-
-    /**
-     * Sets the style width of the element
-     *
-     * ```typescript
-     * // get
-     * let myComboWidth = this.combo.width;
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [width]='250px'></igx-combo>
-     * ```
-     */
-    @HostBinding('style.width')
-    @Input()
-    public width: string;
-
-    /**
-     * @hidden @internal
-     */
-    @HostBinding('class.igx-combo')
-    public cssClass = 'igx-combo'; // Independent of display density, at the time being
-
-    /**
-     * @hidden @internal
-     */
-    @HostBinding(`attr.role`)
-    public role = 'combobox';
-
-    /**
-     * @hidden @internal
-     */
-    @HostBinding('attr.aria-expanded')
-    public get ariaExpanded(): boolean {
-        return !this.dropdown.collapsed;
-    }
-
-    /**
-     * @hidden @internal
-     */
-    @HostBinding('attr.aria-haspopup')
-    public get hasPopUp() {
-        return 'listbox';
-    }
-
-    /**
-     * @hidden @internal
-     */
-    @HostBinding('attr.aria-owns')
-    public get ariaOwns() {
-        return this.dropdown.id;
-    }
-
-    /**
-     * Controls whether custom values can be added to the collection
-     *
-     * ```typescript
-     * // get
-     * let comboAllowsCustomValues = this.combo.allowCustomValues;
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [allowCustomValues]='true'></igx-combo>
+     * <igx-combo [showSearchCaseIcon]='true'></igx-combo>
      * ```
      */
     @Input()
-    public allowCustomValues = false;
+    public showSearchCaseIcon = false;
 
     /**
-     * Configures the drop down list height
-     *
-     * ```typescript
-     * // get
-     * let myComboItemsMaxHeight = this.combo.itemsMaxHeight;
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [itemsMaxHeight]='320'></igx-combo>
-     * ```
+     * An @Input property that controls whether the combo's search box
+     * should be focused after the `opened` event is called
+     * When `false`, the combo's list item container will be focused instead
      */
     @Input()
-    public get itemsMaxHeight(): number {
-        if (this._itemsMaxHeight === null || this._itemsMaxHeight === undefined) {
-            return this.itemHeight * itemsInContainer;
-        }
-        return this._itemsMaxHeight;
-    }
-
-    public set itemsMaxHeight(val: number) {
-        this._itemsMaxHeight = val;
-    }
-
-    /**
-     * Configures the drop down list width
-     *
-     * ```typescript
-     * // get
-     * let myComboItemsWidth = this.combo.itemsWidth;
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [itemsWidth] = '"180px"'></igx-combo>
-     * ```
-     */
-    @Input()
-    public itemsWidth: string;
-
-    /**
-     * Configures the drop down list item height
-     *
-     * ```typescript
-     * // get
-     * let myComboItemHeight = this.combo.itemHeight;
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [itemHeight]='32'></igx-combo>
-     * ```
-     */
-    @Input()
-    public get itemHeight(): number {
-        if (this._itemHeight === null || this._itemHeight === undefined) {
-            return ItemHeights[this.displayDensity];
-        }
-        return this._itemHeight;
-    }
-
-    public set itemHeight(val: number) {
-        this._itemHeight = val;
-    }
-
-    /**
-     * Defines the placeholder value for the combo value field
-     *
-     * ```typescript
-     * // get
-     * let myComboPlaceholder = this.combo.placeholder;
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [placeholder]='newPlaceHolder'></igx-combo>
-     * ```
-     */
-    @Input()
-    public placeholder;
-
-    /**
-     * @hidden @internal
-     */
-    public get inputEmpty(): boolean {
-        return !this.value && !this.placeholder;
-    }
+    public autoFocusSearch = true;
 
     /**
      * Defines the placeholder value for the combo dropdown search field
@@ -665,302 +154,37 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
     @Input()
     public searchPlaceholder = 'Enter a Search Term';
 
-    /**
-     * Defines whether the caseSensitive icon should be shown in the search input
-     *
-     * ```typescript
-     * // get
-     * let myComboShowSearchCaseIcon = this.combo.showSearchCaseIcon;
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [showSearchCaseIcon]='true'></igx-combo>
-     * ```
-     */
-
-    @Input()
-    public showSearchCaseIcon = false;
-
-
-    /**
-     * Combo data source.
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [data]='items'></igx-combo>
-     * ```
-     */
-    @Input()
-    public get data(): any[] | null {
-        return this._data;
-    }
-    public set data(val: any[] | null) {
-        this._data = (val) ? val : [];
-    }
-
-    /**
-     * Combo value data source property.
-     *
-     * ```typescript
-     * // get
-     * let myComboValueKey = this.combo.valueKey;
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [valueKey]='myKey'></igx-combo>
-     * ```
-     */
-    @Input()
-    public valueKey: string = null;
-
-    @Input()
-    public set displayKey(val: string) {
-        this._displayKey = val;
-    }
-
-    /**
-     * Combo text data source property.
-     *
-     * ```typescript
-     * // get
-     * let myComboDisplayKey = this.combo.displayKey;
-     *
-     * // set
-     * this.combo.displayKey = 'val';
-     *
-     * ```
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [displayKey]='myDisplayKey'></igx-combo>
-     * ```
-     */
-    public get displayKey() {
-        return this._displayKey ? this._displayKey : this.valueKey;
-    }
-
-    /**
-     * The item property by which items should be grouped inside the items list. Not usable if data is not of type Object[].
-     *
-     * ```html
-     * <!--set-->
-     * <igx-combo [groupKey]='newGroupKey'></igx-combo>
-     * ```
-     */
-    @Input()
-    public set groupKey(val: string) {
-        this._groupKey = val;
-    }
-
-    /**
-     * The item property by which items should be grouped inside the items list. Not usable if data is not of type Object[].
-     *
-     * ```typescript
-     * // get
-     * let currentGroupKey = this.combo.groupKey;
-     * ```
-     */
-    public get groupKey(): string {
-        return this._groupKey;
-    }
-
-    /**
-     * An @Input property that enabled/disables filtering in the list. The default is `true`.
-     * ```html
-     * <igx-combo [filterable]="false">
-     * ```
-     */
-    @Input()
-    public filterable = true;
-
-    /**
-     * An @Input property that sets groups sorting order.
-     *
-     * @example
-     * ```html
-     * <igx-combo [groupSortingDirection]="groupSortingDirection"></igx-combo>
-     * ```
-     * ```typescript
-     * public groupSortingDirection = SortingDirection.Asc;
-     * ```
-     */
-    @Input()
-    public get groupSortingDirection(): SortingDirection {
-        return this._groupSortingDirection;
-    }
-    public set groupSortingDirection(val: SortingDirection) {
-        this._groupSortingDirection = val;
-    }
-
-    /**
-     * An @Input property that set aria-labelledby attribute
-     * ```html
-     * <igx-combo [ariaLabelledBy]="'label1'">
-     * ```
-     */
-    @Input()
-    @HostBinding('attr.aria-labelledby')
-    public ariaLabelledBy: string;
-
-    /**
-     * An @Input property that enabled/disables combo. The default is `false`.
-     * ```html
-     * <igx-combo [disabled]="'true'">
-     * ```
-     */
-    @Input()
-    public disabled = false;
-
-    /**
-     * An @Input property that sets how the combo will be styled.
-     * The allowed values are `line`, `box`, `border` and `search`. The default is `box`.
-     * ```html
-     * <igx-combo [type]="'line'">
-     * ```
-     */
-    @Input()
-    public get type(): IgxInputGroupType {
-        return this._type || this._inputGroupType || 'box';
-    }
-
-    public set type(val: IgxInputGroupType) {
-        this._type = val;
-    }
-    /**
-     * An @Input property that controls whether the combo's search box
-     * should be focused after the `opened` event is called
-     * When `false`, the combo's list item container will be focused instead
-     */
-    @Input()
-    public autoFocusSearch = true;
-
-    @ViewChild('dropdownItemContainer', { static: true })
-    protected dropdownContainer: ElementRef = null;
-
-    @ViewChild('primitive', { read: TemplateRef, static: true })
-    protected primitiveTemplate: TemplateRef<any>;
-
-    @ViewChild('complex', { read: TemplateRef, static: true })
-    protected complexTemplate: TemplateRef<any>;
-
-    @ViewChild(IgxForOfDirective, { read: IgxForOfDirective, static: true })
-    protected virtDir: IgxForOfDirective<any>;
-
-    /**
-     * Gets if control is valid, when used in a form
-     *
-     * ```typescript
-     * // get
-     * let valid = this.combo.valid;
-     * ```
-     */
-    public get valid(): IgxComboState {
-        return this._valid;
-    }
-
-    /**
-     * Sets if control is valid, when used in a form
-     *
-     * ```typescript
-     * // set
-     * this.combo.valid = IgxComboState.INVALID;
-     * ```
-     */
-    public set valid(valid: IgxComboState) {
-        this._valid = valid;
-        this.comboInput.valid = IgxInputState[IgxComboState[valid]];
-    }
+    /** @hidden @internal */
+    @ViewChild(IgxComboDropDownComponent, { static: true })
+    public dropdown: IgxComboDropDownComponent;
 
     /**
      * @hidden @internal
      */
-    public get searchValue(): string {
-        return this._searchValue;
+    public get inputEmpty(): boolean {
+        return !this.value && !this.placeholder;
     }
 
-    public set searchValue(val: string) {
-        this.filterValue = val;
-        this._searchValue = val;
-    }
-    /**
-     * @hidden @internal
-     */
-    public customValueFlag = true;
-    /**
-     * @hidden @internal
-     */
-    public defaultFallbackGroup = 'Other';
-    /**
-     * @hidden @internal
-     */
-    public filteringOptions: IComboFilteringOptions = {
-        caseSensitive: false
-    };
     /**
      * @hidden @internal
      */
     public filteringLogic = FilteringLogic.Or;
-    /** @hidden @internal */
-    public filterValue = '';
 
     protected stringFilters = IgxStringFilteringOperand;
     protected booleanFilters = IgxBooleanFilteringOperand;
-    protected _groupKey = '';
-    protected _displayKey: string;
     protected _prevInputValue = '';
-    protected _value = '';
-    protected _valid = IgxComboState.INITIAL;
-    protected destroy$ = new Subject<any>();
-    protected _onChangeCallback: (_: any) => void = noop;
-    protected _data = [];
-
-    private _dataType = '';
-    private _searchValue = '';
-    private _type = null;
-    private ngControl: NgControl = null;
-    private _filteredData = [];
-    private _itemHeight = null;
-    private _itemsMaxHeight = null;
-    private _remoteSelection = {};
-    private _onTouchedCallback: () => void = noop;
-    private _overlaySettings: OverlaySettings;
-    private _groupSortingDirection: SortingDirection = SortingDirection.Asc;
 
     constructor(
         protected elementRef: ElementRef,
         protected cdr: ChangeDetectorRef,
-        protected selection: IgxSelectionAPIService,
+        protected selectionService: IgxSelectionAPIService,
         protected comboAPI: IgxComboAPIService,
         protected _iconService: IgxIconService,
         @Optional() @Inject(DisplayDensityToken) protected _displayDensityOptions: IDisplayDensityOptions,
         @Optional() @Inject(IGX_INPUT_GROUP_TYPE) protected _inputGroupType: IgxInputGroupType,
         @Optional() protected _injector: Injector) {
-        super(_displayDensityOptions);
+        super(elementRef, cdr, selectionService, comboAPI, _iconService, _displayDensityOptions, _inputGroupType, _injector);
         this.comboAPI.register(this);
-    }
-
-    /**
-     * @hidden @internal
-     */
-    @HostListener('keydown.ArrowDown', ['$event'])
-    @HostListener('keydown.Alt.ArrowDown', ['$event'])
-    public onArrowDown(event: Event) {
-        event.preventDefault();
-        event.stopPropagation();
-        this.open();
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public onInputClick(event: Event) {
-        event.stopPropagation();
-        event.preventDefault();
-        if (!this.disabled) {
-            this.toggle();
-        }
     }
 
     /**
@@ -987,60 +211,6 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
     }
 
     /**
-     * Gets total count of the virtual data items, when using remote service.
-     *
-     * ```typescript
-     * // get
-     * let count = this.combo.totalItemCount;
-     * ```
-     */
-    public get totalItemCount(): number {
-        return this.virtDir.totalItemCount;
-    }
-    /**
-     * Sets total count of the virtual data items, when using remote service.
-     *
-     * ```typescript
-     * // set
-     * this.combo.totalItemCount(remoteService.count);
-     * ```
-     */
-    public set totalItemCount(count: number) {
-        this.virtDir.totalItemCount = count;
-    }
-
-    /**
-     * The text displayed in the combo input
-     *
-     * ```typescript
-     * // get
-     * let comboValue = this.combo.value;
-     * ```
-     */
-    public get value(): string {
-        return this._value;
-    }
-
-    public set value(val: string) {
-        this._value = val;
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public get filteredData(): any[] | null {
-        return this.filterable ? this._filteredData : this.data;
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public set filteredData(val: any[] | null) {
-        this._filteredData = this.groupKey ? (val || []).filter((e) => e.isHeader !== true) : val;
-        this.checkMatch();
-    }
-
-    /**
      * @hidden @internal
      */
     public handleKeyUp(event: KeyboardEvent): void {
@@ -1051,79 +221,6 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
         } else if (event.key === 'Escape' || event.key === 'Esc') {
             this.toggle();
         }
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public handleKeyDown(event: KeyboardEvent) {
-        if (event.key === 'ArrowUp' || event.key === 'Up') {
-            event.preventDefault();
-            event.stopPropagation();
-            this.close();
-        }
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public handleInputChange(event?: any) {
-        if (event !== undefined) {
-            const args: IComboSearchInputEventArgs = {
-                searchText: typeof event === 'string' ? event : event.target.value,
-                owner: this,
-                cancel: false
-            };
-            this.searchInputUpdate.emit(args);
-            if (args.cancel) {
-                this.filterValue = null;
-            }
-        }
-        this.checkMatch();
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public get dataType(): string {
-        if (this.displayKey) {
-            return DataTypes.COMPLEX;
-        }
-        return DataTypes.PRIMITIVE;
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public get isRemote() {
-        return this.totalItemCount > 0 &&
-            this.valueKey &&
-            this.dataType === DataTypes.COMPLEX;
-    }
-
-    /**
-     * Returns if the specified itemID is selected
-     *
-     * @hidden
-     * @internal
-     */
-    public isItemSelected(item: any): boolean {
-        return this.selection.is_item_selected(this.id, item);
-    }
-
-    /**
-     * Triggers change detection on the combo view
-     */
-    public triggerCheck() {
-        this.cdr.detectChanges();
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public isAddButtonVisible(): boolean {
-        // This should always return a boolean value. If this.searchValue was '', it returns '' instead of false;
-        return this.searchValue !== '' && this.customValueFlag;
     }
 
     /**
@@ -1140,144 +237,12 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
     /**
      * @hidden @internal
      */
-    public addItemToCollection() {
-        if (!this.searchValue) {
-            return;
-        }
-        const newValue = this.searchValue.trim();
-        const addedItem = this.displayKey ? {
-            [this.valueKey]: newValue,
-            [this.displayKey]: newValue
-        } : newValue;
-        if (this.groupKey) {
-            Object.assign(addedItem, { [this.groupKey]: this.defaultFallbackGroup });
-        }
-        // expose shallow copy instead of this.data in event args so this.data can't be mutated
-        const oldCollection = [...this.data];
-        const newCollection = [...this.data, addedItem];
-        const args: IComboItemAdditionEvent = {
-            oldCollection, addedItem, newCollection, owner: this, cancel: false
-        };
-        this.addition.emit(args);
-        if (args.cancel) {
-            return;
-        }
-        this.data.push(args.addedItem);
-        // trigger re-render
-        this.data = cloneArray(this.data);
-        this.select(this.valueKey !== null && this.valueKey !== undefined ?
-            [args.addedItem[this.valueKey]] : [args.addedItem], false);
-        this.customValueFlag = false;
-        this.searchInput?.nativeElement.focus();
-        this.dropdown.focusedItem = null;
-        this.virtDir.scrollTo(0);
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public focusSearchInput(opening?: boolean): void {
-        if (this.displaySearchInput && this.searchInput) {
-            this.searchInput.nativeElement.focus();
-        } else {
-            if (opening) {
-                this.dropdownContainer.nativeElement.focus();
-            } else {
-                this.comboInput.nativeElement.focus();
-                this.toggle();
-            }
-        }
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public onBlur() {
-        if (this.collapsed) {
-            this._onTouchedCallback();
-            if (this.ngControl && this.ngControl.invalid) {
-                this.valid = IgxComboState.INVALID;
-            } else {
-                this.valid = IgxComboState.INITIAL;
-            }
-        }
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public ngOnInit() {
-        this.ngControl = this._injector.get<NgControl>(NgControl, null);
-        const targetElement = this.elementRef.nativeElement;
-        this._overlaySettings = {
-            target: targetElement,
-            scrollStrategy: new AbsoluteScrollStrategy(),
-            positionStrategy: new AutoPositionStrategy(),
-            modal: false,
-            closeOnOutsideClick: true,
-            excludeFromOutsideClick: [targetElement as HTMLElement]
-        };
-        this.selection.set(this.id, new Set());
-        this._iconService.addSvgIconFromText(caseSensitive.name, caseSensitive.value, 'imx-icons');
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public ngAfterViewInit() {
-        this.filteredData = [...this.data];
-
-        if (this.ngControl) {
-            this.ngControl.statusChanges.pipe(takeUntil(this.destroy$)).subscribe(this.onStatusChanged);
-            this.manageRequiredAsterisk();
-            this.cdr.detectChanges();
-        }
-        this.virtDir.chunkPreload.pipe(takeUntil(this.destroy$)).subscribe((e: IForOfState) => {
-            const eventArgs: IForOfState = Object.assign({}, e, { owner: this });
-            this.dataPreLoad.emit(eventArgs);
-        });
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public ngOnDestroy() {
-        this.destroy$.next();
-        this.destroy$.complete();
-        this.comboAPI.clear();
-        this.selection.clear(this.id);
-    }
-
-    /**
-     * @hidden @internal
-     */
     public writeValue(value: any[]): void {
         const selection = Array.isArray(value) ? value : [];
-        const oldSelection = this.selectedItems();
-        this.selection.select_items(this.id, selection, true);
+        const oldSelection = this.selection;
+        this.selectionService.select_items(this.id, selection, true);
         this.cdr.markForCheck();
-        this._value = this.createDisplayText(this.selectedItems(), oldSelection);
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public registerOnChange(fn: any): void {
-        this._onChangeCallback = fn;
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public registerOnTouched(fn: any): void {
-        this._onTouchedCallback = fn;
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public setDisabledState(isDisabled: boolean): void {
-        this.disabled = isDisabled;
+        this._value = this.createDisplayText(this.selection, oldSelection);
     }
 
     /**
@@ -1285,20 +250,6 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
      */
     public getEditElement(): HTMLElement {
         return this.comboInput.nativeElement;
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public get template(): TemplateRef<any> {
-        this._dataType = this.dataType;
-        if (this.itemTemplate) {
-            return this.itemTemplate;
-        }
-        if (this._dataType === DataTypes.COMPLEX) {
-            return this.complexTemplate;
-        }
-        return this.primitiveTemplate;
     }
 
     /**
@@ -1327,68 +278,6 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
     }
 
     /**
-     * A method that opens/closes the combo.
-     *
-     * ```html
-     * <button (click)="combo.toggle()">Toggle Combo</button>
-     * <igx-combo #combo></igx-combo>
-     * ```
-     */
-    public toggle(): void {
-        const overlaySettings = Object.assign({}, this._overlaySettings, this.overlaySettings);
-        this.dropdown.toggle(overlaySettings);
-    }
-
-    /**
-     * A method that opens the combo.
-     *
-     * ```html
-     * <button (click)="combo.open()">Open Combo</button>
-     * <igx-combo #combo></igx-combo>
-     * ```
-     */
-    public open(): void {
-        const overlaySettings = Object.assign({}, this._overlaySettings, this.overlaySettings);
-        this.dropdown.open(overlaySettings);
-    }
-
-    /**
-     * A method that closes the combo.
-     *
-     * ```html
-     * <button (click)="combo.close()">Close Combo</button>
-     * <igx-combo #combo></igx-combo>
-     * ```
-     */
-    public close(): void {
-        this.dropdown.close();
-    }
-
-    /**
-     * Gets drop down state.
-     *
-     * ```typescript
-     * let state = this.combo.collapsed;
-     * ```
-     */
-    public get collapsed(): boolean {
-        return this.dropdown.collapsed;
-    }
-
-    /**
-     * Get current selection state
-     *
-     * @returns Array of selected items
-     * ```typescript
-     * let selectedItems = this.combo.selectedItems();
-     * ```
-     */
-    public selectedItems() {
-        const items = Array.from(this.selection.get(this.id));
-        return items;
-    }
-
-    /**
      * Select defined items
      *
      * @param newItems new items to be selected
@@ -1399,7 +288,7 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
      */
     public select(newItems: Array<any>, clearCurrentSelection?: boolean, event?: Event) {
         if (newItems) {
-            const newSelection = this.selection.add_items(this.id, newItems, clearCurrentSelection);
+            const newSelection = this.selectionService.add_items(this.id, newItems, clearCurrentSelection);
             this.setSelection(newSelection, event);
         }
     }
@@ -1414,7 +303,7 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
      */
     public deselect(items: Array<any>, event?: Event) {
         if (items) {
-            const newSelection = this.selection.delete_items(this.id, items);
+            const newSelection = this.selectionService.delete_items(this.id, items);
             this.setSelection(newSelection, event);
         }
     }
@@ -1428,8 +317,8 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
      * ```
      */
     public selectAllItems(ignoreFilter?: boolean, event?: Event) {
-        const allVisible = this.selection.get_all_ids(ignoreFilter ? this.data : this.filteredData, this.valueKey);
-        const newSelection = this.selection.add_items(this.id, allVisible);
+        const allVisible = this.selectionService.get_all_ids(ignoreFilter ? this.data : this.filteredData, this.valueKey);
+        const newSelection = this.selectionService.add_items(this.id, allVisible);
         this.setSelection(newSelection, event);
     }
 
@@ -1442,9 +331,9 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
      * ```
      */
     public deselectAllItems(ignoreFilter?: boolean, event?: Event): void {
-        let newSelection = this.selection.get_empty();
+        let newSelection = this.selectionService.get_empty();
         if (this.filteredData.length !== this.data.length && !ignoreFilter) {
-            newSelection = this.selection.delete_items(this.id, this.selection.get_all_ids(this.filteredData, this.valueKey));
+            newSelection = this.selectionService.delete_items(this.id, this.selectionService.get_all_ids(this.filteredData, this.valueKey));
         }
         this.setSelection(newSelection, event);
     }
@@ -1478,21 +367,15 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
             this.deselect([itemID], event);
         }
     }
-    /**
-     * Event handlers
-     *
-     * @hidden
-     * @internal
-     */
-    public handleOpening(e: IBaseCancelableBrowserEventArgs) {
-        const args: IBaseCancelableBrowserEventArgs = { owner: this, event: e.event, cancel: e.cancel };
-        this.opening.emit(args);
-        e.cancel = args.cancel;
-    }
 
     /**
      * @hidden @internal
      */
+    public toggleCaseSensitive() {
+        this.filteringOptions = { caseSensitive: !this.filteringOptions.caseSensitive };
+    }
+
+    /** @hidden @internal */
     public handleOpened() {
         this.triggerCheck();
 
@@ -1507,41 +390,11 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
         this.opened.emit({ owner: this });
     }
 
-    /**
-     * @hidden @internal
-     */
-    public handleClosing(e: IBaseCancelableBrowserEventArgs) {
-        const args: IBaseCancelableBrowserEventArgs = { owner: this, event: e.event, cancel: e.cancel };
-        this.closing.emit(args);
-        e.cancel = args.cancel;
-        if (e.cancel) {
-            return;
-        }
-        this.searchValue = '';
-        if (!e.event) {
-            this.comboInput?.nativeElement.focus();
-        }
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public handleClosed() {
-        this.closed.emit({ owner: this });
-    }
-
-    /**
-     * @hidden @internal
-     */
-    public toggleCaseSensitive() {
-        this.filteringOptions = { caseSensitive: !this.filteringOptions.caseSensitive };
-    }
-
     protected setSelection(newSelection: Set<any>, event?: Event): void {
-        const removed = diffInSets(this.selection.get(this.id), newSelection);
-        const added = diffInSets(newSelection, this.selection.get(this.id));
+        const removed = diffInSets(this.selectionService.get(this.id), newSelection);
+        const added = diffInSets(newSelection, this.selectionService.get(this.id));
         const newSelectionAsArray = Array.from(newSelection);
-        const oldSelectionAsArray = Array.from(this.selection.get(this.id) || []);
+        const oldSelectionAsArray = Array.from(this.selectionService.get(this.id) || []);
         const displayText = this.createDisplayText(newSelectionAsArray, oldSelectionAsArray);
         const args: IComboSelectionChangingEventArgs = {
             newSelection: newSelectionAsArray,
@@ -1555,7 +408,7 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
         };
         this.selectionChanging.emit(args);
         if (!args.cancel) {
-            this.selection.select_items(this.id, args.newSelection, true);
+            this.selectionService.select_items(this.id, args.newSelection, true);
             if (displayText !== args.displayText) {
                 this._value = args.displayText;
             } else {
@@ -1563,104 +416,6 @@ export class IgxComboComponent extends DisplayDensityBase implements IgxComboBas
             }
             this._onChangeCallback(args.newSelection);
         }
-    }
-
-    protected onStatusChanged = () => {
-        if ((this.ngControl.control.touched || this.ngControl.control.dirty) &&
-            (this.ngControl.control.validator || this.ngControl.control.asyncValidator)) {
-            if (!this.collapsed || this.inputGroup.isFocused) {
-                this.valid = this.ngControl.invalid ? IgxComboState.INVALID : IgxComboState.VALID;
-            } else {
-                this.valid = this.ngControl.invalid ? IgxComboState.INVALID : IgxComboState.INITIAL;
-            }
-        } else {
-            // B.P. 18 May 2021: IgxDatePicker does not reset its state upon resetForm #9526
-            this.valid = IgxComboState.INITIAL;
-        }
-        this.manageRequiredAsterisk();
-    };
-
-    protected manageRequiredAsterisk(): void {
-        if (this.ngControl && this.ngControl.control.validator) {
-            // Run the validation with empty object to check if required is enabled.
-            const error = this.ngControl.control.validator({} as AbstractControl);
-            this.inputGroup.isRequired = error && error.required;
-        }
-    }
-
-    /**
-     * Constructs the combo display value
-     * If remote, caches the key displayText
-     * If not, just combine the object.displayKeys
-     */
-    protected createDisplayText(newSelection: any[], oldSelection: any[]) {
-        let value = '';
-        if (this.isRemote) {
-            if (newSelection.length) {
-                const removedItems = oldSelection.filter(e => newSelection.indexOf(e) < 0);
-                const addedItems = newSelection.filter(e => oldSelection.indexOf(e) < 0);
-                this.registerRemoteEntries(addedItems);
-                this.registerRemoteEntries(removedItems, false);
-                value = Object.keys(this._remoteSelection).map(e => this._remoteSelection[e]).join(', ');
-            } else {
-                // If new selection is empty, clear all items
-                this.registerRemoteEntries(oldSelection, false);
-            }
-        } else {
-            value = this.concatDisplayText(newSelection);
-        }
-        return value;
-    }
-
-    protected checkMatch(): void {
-        const itemMatch = this.filteredData.some(this.findMatch);
-        this.customValueFlag = this.allowCustomValues && !itemMatch;
-    }
-
-    protected findMatch = (element: any): boolean => {
-        const value = this.displayKey ? element[this.displayKey] : element;
-        return value.toString().toLowerCase() === this.searchValue.trim().toLowerCase();
-    };
-
-    /** Contains key-value pairs of the selected valueKeys and their resp. displayKeys */
-    private registerRemoteEntries(ids: any[], add = true) {
-        if (add) {
-            const selection = this.getValueDisplayPairs(ids);
-            for (const entry of selection) {
-                this._remoteSelection[entry[this.valueKey]] = entry[this.displayKey];
-            }
-        } else {
-            for (const entry of ids) {
-                delete this._remoteSelection[entry];
-            }
-        }
-    }
-
-    /**
-     * For `id: any[]` returns a mapped `{ [combo.valueKey]: any, [combo.displayKey]: any }[]`
-     */
-    private getValueDisplayPairs(ids: any[]) {
-        return this.data.filter(entry => ids.indexOf(entry[this.valueKey]) > -1).map(e => ({
-            [this.valueKey]: e[this.valueKey],
-            [this.displayKey]: e[this.displayKey]
-        }));
-    }
-
-    /** Returns a string that should be populated in the combo's text box */
-    private concatDisplayText(selection: any[]): string {
-        const value = this.displayKey !== null && this.displayKey !== undefined ?
-            this.convertKeysToItems(selection).map(entry => entry[this.displayKey]).join(', ') :
-            selection.join(', ');
-        return value;
-    }
-
-    /** if there is a valueKey - map the keys to data items, else - just return the keys */
-    private convertKeysToItems(keys: any[]) {
-        if (this.comboAPI.valueKey === null) {
-            return keys;
-        }
-        // map keys vs. filter data to retain the order of the selected items
-        return keys.map(key => this.data.find(entry => entry[this.valueKey] === key)).filter(e => e !== undefined);
     }
 }
