+++ projects/igniteui-angular/src/lib/grids/grid/grid.nested.props.spec.ts
@@ -545,7 +545,7 @@ describe('Edit cell with data of type Array #grid', () => {
         expect(grid.cellEditEnter.emit).toHaveBeenCalledTimes(1);
         expect(grid.cellEditEnter.emit).toHaveBeenCalledWith(cellArgs);
         expect(cell.editMode).toBeTruthy();
-        expect(combo.selectedItems().length).toEqual(3);
+        expect(combo.selection.length).toEqual(3);
 
         combo.deselect([cell.editValue[0], cell.editValue[1]]);
         fixture.detectChanges();
@@ -607,7 +607,7 @@ describe('Edit cell with data of type Array #grid', () => {
         expect(grid.cellEditEnter.emit).toHaveBeenCalledTimes(1);
         expect(grid.cellEditEnter.emit).toHaveBeenCalledWith(cellArgs);
         expect(cell.editMode).toBeTruthy();
-        expect(combo.selectedItems().length).toEqual(3);
+        expect(combo.selection.length).toEqual(3);
 
         combo.deselect([cell.editValue[0], cell.editValue[1]]);
         fixture.detectChanges();
@@ -675,7 +675,7 @@ describe('Edit cell with data of type Array #grid', () => {
         expect(grid.rowEditEnter.emit).toHaveBeenCalledWith(rowArgs);
         expect(row.inEditMode).toBeTruthy();
         expect(cell.editMode).toBeTruthy();
-        expect(combo.selectedItems().length).toEqual(3);
+        expect(combo.selection.length).toEqual(3);
 
         combo.deselect([cell.editValue[0], cell.editValue[1]]);
         fixture.detectChanges();
@@ -738,7 +738,7 @@ describe('Edit cell with data of type Array #grid', () => {
         expect(grid.rowEditEnter.emit).toHaveBeenCalledWith(rowArgs);
         expect(row.inEditMode).toBeTruthy();
         expect(cell.editMode).toBeTruthy();
-        expect(combo.selectedItems().length).toEqual(3);
+        expect(combo.selection.length).toEqual(3);
 
         combo.deselect([cell.editValue[0], cell.editValue[1]]);
         fixture.detectChanges();
